backend-ml/requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
python-multipart
httpx
psycopg2-binary
twilio
python-dotenv
